requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Limiter==3.12
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dotenv==1.0.1
redis==5.2.1
rich==13.9.4
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
